Modificações no artigo e desenvolvimento da implementação da estrutura da árvore

+++ include/random.h
@@ -2,6 +2,6 @@
 #define __RANDOM_H__
 
 int randomInt( int min, int max );
-int randomFloat( float min, float max );
+float randomFloat( float min, float max );
 
 #endif
